ファイルダウンロード処理を行うクラスを追加#15
