+++ src/fm-client/abilities/inventory/inventory/index.js
@@ -4,10 +4,7 @@ const MenuItem = require('../../../components/MenuItem')
 const Native = require('../../../components/Native');
 const ButtonSymbol = require('../../../components/ButtonSymbol');
 const RequestHijacking = require('../../../components/RequestHijacking');
-
-const https = require("https");
-const fs = require('fs');
-const path = require('path');
+const FileDownloader = require('../../../components/FileDownloader');
 
 const INDEX_BUTTON = 2
 const MENU_ITEM = new MenuItem(10, 1, 1)
@@ -47,7 +44,7 @@ module.exports = class Inventory extends AbstractSinglePage {
       // ファイルをダウンロード
       const rh = new RequestHijacking(page)
       await rh.intercept('/JMODE_ASP/SlipList$',
-        Inventory.sendRequest_,
+        FileDownloader.request,
         options
       )
       await page.click('.excelDLDiv input[name=smt]');
@@ -59,39 +56,5 @@ module.exports = class Inventory extends AbstractSinglePage {
         statusText: e.message//'原因不明のエラー',
       }
     }
-  }
-
-  static async sendRequest_(request, cookies, userData) {
-    const options = {
-      encoding: null,
-      method: request._method,
-      headers: request._headers,
-      resolveWithFullResponse: true,
-    }
-    options.headers.Cookie = cookies.map(cookie => cookie.name + '=' + cookie.value).join(';');
-  
-    return new Promise((resolve, reject) => {
-      const req = https.request(request._url, options, (response) => {
-        if(response.statusCode !== 200) {
-          reject()
-          return
-        }
-        
-        const machedArray = RegExp('filename=(.*)', 'g').exec(response.headers['content-disposition'])
-        const filepath = path.join(userData.directoryToSaveFile || '', machedArray[1])
-
-        const stream = fs.createWriteStream(filepath, { encoding: null });
-        response.on('data', (chunk) => stream.write(chunk));
-        response.on('end', () => {
-          stream.end();
-          request.abort()
-          resolve()
-        });
-      });
-      
-      req.on('error', (e) => reject(e));
-      req.write(request._postData);
-      req.end();
-    });
   }
 }
